# docs/api-docs.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Untripical API",
        "description": "API aplication Untripical",
        "contact": {
            "name": "Mateusz",
            "email": "mateusz@untripical.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Untripical server"
        }
    ],
    "paths": {
...
        "/review/place/{placeId}/{orderIndex}": {
            "put": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "updatePlaceReviewWithOrderIndex",
                "parameters": [
                    {
                        "name": "orderIndex",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "placeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRequestPlaceDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "deletePlaceReviewWithOrderIndex",
                "parameters": [
                    {
                        "name": "orderIndex",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "placeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/review/guideDetails/{guideDetailsId}/{orderIndex}": {
            "put": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "updateGuideDetailsReviewWithOrderIndex",
                "parameters": [
                    {
                        "name": "orderIndex",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "guideDetailsId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRequestGuideDetailsDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewResponseGuideDetailsDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "deleteGuideDetailsReviewWithOrderIndex",
                "parameters": [
                    {
                        "name": "orderIndex",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "guideDetailsId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/place/{verificationStatus}/{placeId}": {
            "put": {
                "tags": [
                    "place-controller"
                ],
                "operationId": "setVerificationStatusOfPlaceByGuide",
                "parameters": [
                    {
                        "name": "verificationStatus",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "WAITING_FOR_APPROVAL",
                                "APPROVED",
                                "REJECTED"
                            ]
                        }
                    },
                    {
                        "name": "placeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlaceResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/place/{placeId}": {
            "put": {
                "tags": [
                    "place-controller"
                ],
                "operationId": "updatePlace",
                "parameters": [
                    {
                        "name": "placeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlaceUpdateDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlaceResponseDTO"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "place-controller"
                ],
                "operationId": "deletePlaceById",
                "parameters": [
                    {
                        "name": "placeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlaceResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/review/place": {
            "post": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "addPlaceReview",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewRequestPlaceDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewResponsePlaceDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/review/guideDetails": {
            "post": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "addGuideDetailsReview",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ReviewRequestGuideDetailsDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/ReviewResponseGuideDetailsDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/place": {
            "post": {
                "tags": [
                    "place-controller"
                ],
                "operationId": "addTripPlace",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlaceRequestDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/PlaceResponseDTO"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/review/reviews/created-for-place/{placeId}": {
            "get": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "numberOfReviewsForOnePlace",
                "parameters": [
                    {
                        "name": "placeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/review/reviews/created-for-guide/{guideDetailsId}": {
            "get": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "numberOfReviewsCreatedForOneGuideDetails",
                "parameters": [
                    {
                        "name": "guideDetailsId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "integer",
                                    "format": "int32"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/review/place/{placeId}": {
            "get": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "getAllPlaceReviewsWithId",
                "parameters": [
                    {
                        "name": "placeId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponsePlaceDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/review/place/numberOfStars/{numberOfStars}": {
            "get": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "getReviewsForPlaceByNumberOfStars",
                "parameters": [
                    {
                        "name": "numberOfStars",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponsePlaceDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/review/place/containing/{keyword}": {
            "get": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "getPlaceReviewsContainingKeyword",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponsePlaceDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/review/guideDetails/{guideId}": {
            "get": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "getAllGuideDetailsReviews",
                "parameters": [
                    {
                        "name": "guideId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponseGuideDetailsDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/review/guideDetails/numberOfStars/{numberOfStars}": {
            "get": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "getReviewsForGuideDetailsByNumberOfStars",
                "parameters": [
                    {
                        "name": "numberOfStars",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "double"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponseGuideDetailsDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/review/guideDetails/containing/{keyword}": {
            "get": {
                "tags": [
                    "review-controller"
                ],
                "operationId": "getGuideDetailsReviewsContainingKeyword",
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ReviewResponseGuideDetailsDTO"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateRequestPlaceDTO": {
                "type": "object",
                "properties": {
                    "numberOfStars": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
            "UpdateRequestGuideDetailsDTO": {
                "type": "object",
                "properties": {
                    "numberOfStars": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    }
                }
            },
            "ReviewResponseGuideDetailsDTO": {
                "type": "object",
                "properties": {
                    "numberOfStars": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "guideId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "usernameWhoWroteReview": {
                        "type": "string"
                    },
                    "orderIndex": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "PlaceResponseDTO": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "addressStreet": {
                        "type": "string"
                    },
                    "addressNumber": {
                        "type": "string"
                    },
                    "placeType": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string"
                    },
                    "isActive": {
                        "type": "boolean"
                    },
                    "photoUrl": {
                        "type": "string"
                    },
                    "verificationStatus": {
                        "type": "string",
                        "enum": [
                            "WAITING_FOR_APPROVAL",
                            "APPROVED",
                            "REJECTED"
                        ]
                    },
                    "regionId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "userId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "announcements": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AnnouncementResponseDTO"
                        }
                    },
                    "reviews": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ReviewResponsePlaceDTO"
                        }
                    }
                }
            },
...
            "ReviewResponsePlaceDTO": {
                "type": "object",
                "properties": {
                    "numberOfStars": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "placeId": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "usernameWhoWroteReview": {
                        "type": "string"
                    },
                    "orderIndex": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "PlaceUpdateDTO": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "city": {
                        "type": "string"
                    },
                    "addressStreet": {
                        "type": "string"
                    },
                    "addressNumber": {
                        "type": "string"
                    },
                    "postalCode": {
                        "type": "string"
                    },
                    "placeType": {
                        "type": "string"
                    },
                    "photoUrl": {
                        "type": "string"
                    },
                    "regionId": {
                        "type": "integer",
                        "format": "int64"
                    }
                }
            },
...
            "ReviewRequestPlaceDTO": {
                "type": "object",
                "properties": {
                    "numberOfStars": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string",
                        "minLength": 1
                    },
                    "placeId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "numberOfStars",
                    "placeId"
                ]
            },
            "ReviewRequestGuideDetailsDTO": {
                "type": "object",
                "properties": {
                    "numberOfStars": {
                        "type": "number",
                        "format": "double"
                    },
                    "description": {
                        "type": "string",
                        "minLength": 1
                    },
                    "guideId": {
                        "type": "integer",
                        "format": "int64"
                    }
                },
                "required": [
                    "guideId",
                    "numberOfStars"
                ]
            },
...
        }
    }
}